requirements.txt
flake8==3.3.0
flake8-docstrings==1.1.0
django-debug-toolbar==1.7 # pyup: <1.8
freezegun==0.3.9
isort==4.2.15
pytest-django==3.1.2
pytest-cov==2.5.1
pytest-factoryboy==1.3.1
factory-boy==2.8.1
Faker==0.7.18
transifex-client==0.12.4
#https://django-autoslug.readthedocs.io/en/latest/fields.html